move TrustedRandomAccess into its own module
